tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-dispose-ordering.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  let oncurrentchange_called = false;
  let ondispose_called = false;

  let original_entry = appHistory.current;
  original_entry.ondispose = t.step_func(() => {
    assert_true(oncurrentchange_called);
    ondispose_called = true;
  });

  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_called = true;
    assert_equals(e.from, original_entry);
    assert_equals(e.from.index, -1);
    assert_equals(e.navigationType, "replace");
    assert_equals(appHistory.current.index, 0);
  });
  appHistory.navigate("#foo", { replace: true });
  assert_true(oncurrentchange_called);
  assert_true(ondispose_called);
}, "Ordering between AppHistoryCurrentChangeEvent and AppHistoryEntry dispose events");
</script>
